requirements.txt
aiogram==2.9.2
aiohttp==3.6.2
aioredis==1.3.1
python-dotenv==0.14.0
gino~=1.0.1
sqlalchemy~=1.3.19
